+++ Luna/src/vm/value.c
@@ -3,8 +3,6 @@
 #include <stdio.h>
 #include "memory.h"
 #include "vm.h"
-
-
 // --- 扩容逻辑 (Cold Path) ---
 // [修改] 增加 VM* vm 参数
 #if defined(_MSC_VER)
@@ -15,22 +13,20 @@ __attribute__((noinline))
 void growValueArray(VM* vm, ValueArray* array) {
     u32 oldCapacity = array->capacity;
     u32 newCapacity = GROW_CAPACITY(oldCapacity);
-   
+  
     // [修改] 使用传入的 vm 调用内存分配器
     array->values = GROW_ARRAY(vm, Value, array->values, oldCapacity, newCapacity);
-   
+  
     array->capacity = newCapacity;
 }
-
 // [修改] 增加 VM* vm 参数
 void freeValueArray(VM* vm, ValueArray* array) {
     // [修改] 使用传入的 vm 释放内存
     FREE_ARRAY(vm, Value, array->values, array->capacity);
-   
+  
     // 重置元数据
     initValueArray(array);
 }
-
 void printValue(Value value) {
 #ifdef NAN_BOXING
     if (IS_NUMBER(value)) {
@@ -43,7 +39,7 @@ void printValue(Value value) {
         printf("nil");
     }
     else {
-        printf("<BAD VALUE: 0x%llx>", (unsigned long long)value);
+        printf("<BAD VALUE: 0x%llx>", (u64)value);
     }
 #else
     switch (value.type) {
